ordenes de atencion y agenda casi listo

+++ src/main/java/com/chiletel/dto/DañoNuevoDTO.java
@@ -1,12 +1,13 @@
 package com.chiletel.dto;
 
+import java.math.BigInteger;
 import java.util.Date;
 import io.swagger.annotations.ApiModelProperty;
 import lombok.Data;
 
 @Data
 public class DañoNuevoDTO {
-	private int cliente;
+	private BigInteger cliente;
 	private String tipoDaño;
 	private String descripcion;
